Moved widget functions to a TunableWidget class
